Move backdrop into own component
